feat(backend): scores upsert/submit + публичный leaderboard
PUT /teams/{id}/scores — батч-апсерт черновиков, валидация что value в [0, max_score].
POST /teams/{id}/scores/submit — финализация, требует оценки по всем критериям события.
GET /events/{id}/leaderboard — взвешенный рейтинг, среднее по судьям, partial-судьи игнорируются.

+++ backend/app/models/score.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import enum
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     DateTime,
@@ -40,7 +41,7 @@ class Score(Base):
     criterion_id: Mapped[uuid.UUID] = mapped_column(
         Uuid(), ForeignKey("criteria.id", ondelete="CASCADE"), nullable=False, index=True
     )
-    value: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False)
+    value: Mapped[Decimal] = mapped_column(Numeric(5, 2), nullable=False)
     comment: Mapped[str | None] = mapped_column(String(2000), nullable=True)
     status: Mapped[ScoreStatus] = mapped_column(
         Enum(ScoreStatus, name="score_status"), nullable=False, default=ScoreStatus.draft
